+++ releases.md
@@ -1,5 +1,9 @@
 # Releases
 
+## Unreleased
+
+  - Add `async:container:supervisor:reload` command to restart the container (blue/green deployment).
+
 ## v0.1.0
 
   - Initial implementation.